DAT-1175 refactor search
