+++ cmdb/updater/versions/updater_20200512.py
@@ -91,4 +91,4 @@ class Update20200512(Updater):
 
     def __clear_up_types(self):
         """Removes the category_id field from type collection"""
-        self.dbm.unset_update_many(collection=TypeModel.COLLECTION, filter={}, data='category_id')
+        self.dbm.unset_update_many(collection=TypeModel.COLLECTION, criteria={}, data='category_id')
